+++ NEWS.md
@@ -1,3 +1,13 @@
+# Unreleased
+
+Bug fixes:
+
+* Fixed bug in `FeatureMatrix()` when cells information not present in Fragment object ([#803](https://github.com/timoast/signac/issues/803))
+
+Other changes:
+
+* Added `idf` parameter to `RunTFIDF()` to use precomputed IDF vector
+
 # Signac 1.4.0
 
 New functionality: